requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
blinker==1.9.0
click==8.1.8
discord.py==2.5.2
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.2.0
propcache==0.3.1
python-dotenv==1.1.0
typing_extensions==4.13.0
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
